solid_mechanics -> displacement for variable filenames
This naming was stupid and inconsistent so we'll change it now
while refactoring. We can have different kinds of solid mechanics
variables, so let's make the name displacement since that's what
they actually are. This is file renaming only. Class renaming
will come subsequently.
