+++ src/Makefile.am
@@ -191,8 +191,8 @@ libgrins_la_SOURCES += variables/src/primitive_flow_variables.C
 libgrins_la_SOURCES += variables/src/primitive_flow_fe_variables.C
 libgrins_la_SOURCES += variables/src/primitive_temp_variables.C
 libgrins_la_SOURCES += variables/src/primitive_temp_fe_variables.C
-libgrins_la_SOURCES += variables/src/solid_mechanics_variables.C
-libgrins_la_SOURCES += variables/src/solid_mechanics_fe_variables.C
+libgrins_la_SOURCES += variables/src/displacement_variables.C
+libgrins_la_SOURCES += variables/src/displacement_fe_variables.C
 libgrins_la_SOURCES += variables/src/species_mass_fracs_variables.C
 libgrins_la_SOURCES += variables/src/species_mass_fracs_fe_variables.C
 libgrins_la_SOURCES += variables/src/thermo_pressure_variable.C
@@ -425,8 +425,8 @@ include_HEADERS += variables/include/grins/primitive_flow_variables.h
 include_HEADERS += variables/include/grins/primitive_flow_fe_variables.h
 include_HEADERS += variables/include/grins/primitive_temp_variables.h
 include_HEADERS += variables/include/grins/primitive_temp_fe_variables.h
-include_HEADERS += variables/include/grins/solid_mechanics_variables.h
-include_HEADERS += variables/include/grins/solid_mechanics_fe_variables.h
+include_HEADERS += variables/include/grins/displacement_variables.h
+include_HEADERS += variables/include/grins/displacement_fe_variables.h
 include_HEADERS += variables/include/grins/species_mass_fracs_variables.h
 include_HEADERS += variables/include/grins/species_mass_fracs_fe_variables.h
 include_HEADERS += variables/include/grins/thermo_pressure_variable.h
